First version.
This version retrives the password from the current page and
apply zxcvbn function to estimate the password strength.
Application handles 3 different error conditions -

1. The page doesn't have the password field
2. The password field is empty
3. The content script has returned an empty response.

After installation or update a notice is shown to the user to
reload the tab or restart the chrome.

The password strength range is 0-4 and meter shows it in
different colors.