Stopped automatic retrival of password.
The first version use to automatically retrive password from
the current page when the popup was opened. From a user
perspective this can be a conern. Also no facility was provided
for the user to enter the password.

This version solves these 2 problems. Now user has to explicitly
hit on the "Estimate Strength" button if he wishes to check the
password strength.